gcp/function/src/requirements.txt
Flask==2.1.2
functions-framework==3.1.0
google-cloud-bigquery==3.2.0
